+++ appconfigservice/app/appconfig/db.py
@@ -2,22 +2,19 @@ from flask import current_app, g
 from pymongo.mongo_client import MongoClient
 import pymongo
 
+client = None
 
 def get_db():
     if 'db' not in g:
-        g.client = MongoClient(current_app.config['APP_CONFIG_MONGO_URL'])
-        g.db = g.client.get_database(name=current_app.config['APP_CONFIG_DB_NAME'])
-        g.collection = g.db[current_app.config['APP_CONFIGS_COLLECTION']]
-        g.collection.create_index([("mobileAppVersion", pymongo.DESCENDING)], unique=True)
+        g.db = client.get_database(name=current_app.config['APP_CONFIG_DB_NAME'])
     return g.db
 
 
-def close_db(e=None):
-    client = g.pop('client', None)
+def init_db(app):
+    global client
+    client = MongoClient(app.config['APP_CONFIG_MONGO_URL'])
 
-    if client is not None:
-        client.close()
-
-
-def init_app(app):
-    app.teardown_appcontext(close_db)
+    # Create indexes on app start
+    db = client.get_database(name=app.config['APP_CONFIG_DB_NAME'])
+    app_configs = db[app.config['APP_CONFIGS_COLLECTION']]
+    app_configs.create_index([("mobileAppVersion", pymongo.DESCENDING)], unique=True)