+++ eventservice/gunicorn.config.py
@@ -3,3 +3,5 @@
 bind = '0.0.0.0:5000'
 
 workers = 4
+worker_class = 'gevent'
+worker_connections = 1000