+++ loggingservice/__init__.py
@@ -8,6 +8,6 @@ def create_app():
     app.url_map.strict_slashes = False
     app.register_blueprint(logging_rest_service.bp)
     app.config.from_pyfile('config.py', silent=True)
-    db.init_app(app)
+    db.init_db(app)
 
     return app