+++ loggingservice/db.py
@@ -1,20 +1,16 @@
 from flask import current_app, g
 from pymongo.mongo_client import MongoClient
 
+client = None
 
 def get_db():
     if 'db' not in g:
-        g.client = MongoClient(current_app.config['LOGGING_MONGO_URL'])
-        g.db = g.client.get_database(name=current_app.config['LOGGING_DB_NAME'])
+        g.db = client.get_database(name=current_app.config['LOGGING_DB_NAME'])
     return g.db
 
 
-def close_db(e=None):
-    client = g.pop('client', None)
+def init_db(app):
+    global client
 
-    if client is not None:
-        client.close()
-
-
-def init_app(app):
-    app.teardown_appcontext(close_db)
+    if app.config.get('LOGGING_MONGO_URL', None):
+        client = MongoClient(app.config['LOGGING_MONGO_URL'])